% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthBackend.R
\docType{data}
\name{AuthBackend}
\alias{AuthBackend}
\title{Create AuthBackend}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Creates AuthBackend class object.
}
\section{Construction}{


Note: This object is typically constructed via a derived classes, e.g.
\link{AuthBackendBasic} or \link{AuthBackendBearer}.\preformatted{AuthBackend$new(FUN, auth_header_prefix)
}
\itemize{
\item \code{FUN} :: \code{function}\cr
Authentication handler function.
\item \code{auth_header_prefix} :: \code{character(1)}\cr
Authentication HTTP header prefix.
}
}

\section{Methods}{

\itemize{
\item \code{authenticate()}\cr
\code{simpleError}\cr
This placeholder. It must be implemented in the subclass.
}
}

\references{
\href{https://github.com/loanzen/falcon-auth}{falcon-auth}
}
\seealso{
Other AuthBackend: \code{\link{AuthBackendBasic}},
  \code{\link{AuthBackendBearer}},
  \code{\link{AuthMiddleware}}
}
\concept{AuthBackend}
\keyword{internal}
