% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthBackendBasic.R
\docType{data}
\name{AuthBackendBasic}
\alias{AuthBackendBasic}
\alias{BasicAuthBackend}
\title{Basic authorization backend}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Creates AuthBackendBasic class object.
}
\section{Construction}{
\preformatted{AuthBackendBasic$new(FUN)
}
\itemize{
\item \code{FUN} :: \code{function}\cr
Function to perform authentication.
}
}

\section{Methods}{

\itemize{
\item \code{authenticate(request, response)}\cr
\link{RestRserveRequest}, \link{RestRserveResponse} -> \code{NULL}\cr
Provide authentication for the given request.
}
}

\examples{
# init users database
user_db = list(
  "user-1" = "password-1",
  "user-2" = "password-2"
)
# define authentication handler
auth_fun = function(user, password) {
  if (is.null(user_db[[user]])) return(FALSE) # not found
  if (!identical(user_db[[user]], password)) return(FALSE) # incorrect
  return(TRUE)
}
# init backend
auth_backend = AuthBackendBasic$new(FUN = auth_fun)

# test backend
# define credentials (see RFC)
creds = jsonlite::base64_enc("user-1:password-1")
# generate request headers
h = sprintf("Authorization: Basic \%s", creds)
# simulate request
rq = RestRserveRequest$new(path = "/", headers = h)
# init response object
rs = RestRserveResponse$new()
# perform authentication
auth_backend$authenticate(rq, rs) # TRUE

}
\references{
\href{https://tools.ietf.org/html/rfc7617}{RFC7617}
\href{https://en.wikipedia.org/wiki/Basic_access_authentication}{Wikipedia}
}
\seealso{
\link{RestRserveAuthMiddleware} \link{RestRserveRequest} \link{RestRserveResponse}

Other AuthBackend: \code{\link{AuthBackendBearer}},
  \code{\link{AuthBackend}},
  \code{\link{RestRserveAuthMiddleware}}
}
\concept{AuthBackend}
\keyword{datasets}
