% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backend.R
\name{Backend}
\alias{Backend}
\title{Creates Backend object}
\description{
Creates Backend object.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Backend-new}{\code{Backend$new()}}
\item \href{#method-Backend-start}{\code{Backend$start()}}
\item \href{#method-Backend-clone}{\code{Backend$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Backend-new"></a>}}
\if{latex}{\out{\hypertarget{method-Backend-new}{}}}
\subsection{Method \code{new()}}{
Creates Backend object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Backend$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Backend-start"></a>}}
\if{latex}{\out{\hypertarget{method-Backend-start}{}}}
\subsection{Method \code{start()}}{
Starts backend.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Backend$start(app, port, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{app}}{\link{Application} object.}

\item{\code{port}}{HTTP port.}

\item{\code{...}}{Passed to backend.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Backend-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Backend-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Backend$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
