% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestRserveAuthMiddleware.R
\docType{data}
\name{BearerAuthBackend}
\alias{BearerAuthBackend}
\title{Bearer token authorization backend}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BearerAuthBackend
}
\description{
\url{https://swagger.io/docs/specification/authentication/bearer-authentication/}
}
\keyword{datasets}
