% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\docType{data}
\name{constants}
\alias{constants}
\alias{IGNORE}
\alias{ERROR}
\alias{WARNING}
\alias{INFO}
\alias{DEBUG}
\alias{TRACE}
\title{log level constants}
\format{An object of class \code{integer} of length 1.}
\usage{
IGNORE

ERROR

WARNING

INFO

DEBUG

TRACE
}
\description{
log level constants
}
\keyword{datasets}
