% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentHandlers.R
\docType{data}
\name{ContentHandlers}
\alias{ContentHandlers}
\title{Controls how RestRserve encodes and decodes different content types}
\format{An object of class \code{RestRserveContentHandler} (inherits from \code{R6}) of length 9.}
\usage{
ContentHandlers
}
\description{
Controls how RestRserve encodes and decodes different content types
}
\keyword{datasets}
