% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentHandlers.R
\docType{data}
\name{ContentHandlersFactory}
\alias{ContentHandlersFactory}
\title{Content handlers collection}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Controls how RestRserve encodes and decodes different content types.
}
\section{Construction}{
\preformatted{ContentHandlersFactory$new()
}
}

\section{Fields}{

\itemize{
\item \code{handlers} :: \code{environment()}\cr
Handlers storage environment.
}
}

\section{Methods}{

\itemize{
\item \code{get_encode(content_type)}\cr
\code{character(1)} -> \code{function}\cr
Get encoder function for the specific content type.
\item \code{get_decode(content_type)}\cr
\code{character(1)} -> \code{function}\cr
Get decoder function for the specific content type.
\item \code{set_encode(content_type, FUN)}\cr
\code{character(1)}, \code{function} -> \code{self}\cr
Set hahndler to encode body for the specific content type.
\item \code{set_decode(content_type, FUN)}\cr
\code{character(1)}, \code{function} -> \code{self}\cr
Set hahndler to decode body for the specific content type.
}
}

\keyword{internal}
