% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{create_response}
\alias{create_response}
\title{creates http response}
\usage{
create_response(payload = "", content_type = "text/html",
  headers = character(0), status_code = 200L)
}
\arguments{
\item{payload}{can be a character vector of length one or a raw vector.}

\item{content_type}{must be a character vector of length one}

\item{headers}{must be a character vector - the elements will have CRLF appended.
Neither Content-type nor Content-length may be used.}

\item{status_code}{must be an integer (default is 200L)}
}
