% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContentHandlers.R
\docType{data}
\name{EncodeDecodeMiddleware}
\alias{EncodeDecodeMiddleware}
\title{Creates EncodeDecodeMiddleware middleware object}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Controls how RestRserve encodes and decodes different content types.
\strong{This middleware is passed by default to the \link{Application} constructor}.
Designed to work jointly with \link{ContentHandlers}
This class inherits from \link{Middleware}.
}
\section{Construction}{
\preformatted{EncodeDecodeMiddleware$new(id = "EncodeDecodeMiddleware")
}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Middleware id
}
}

\seealso{
\link{Middleware} \link{Application} \link{ContentHandlers}
}
\keyword{datasets}
