% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_date.R
\name{to_http_date}
\alias{to_http_date}
\alias{from_http_date}
\title{Dealing with HTTP Dates}
\usage{
to_http_date(dtm)

from_http_date(str)
}
\arguments{
\item{dtm}{POSIXct object.}

\item{str}{Character string.}
}
\description{
Convert POSIXct to HTTP Date and vice versa.
}
\examples{
# convert POSIXct to HTTP date string
to_http_date(0) # Thu, 01 Jan 1970 00:00:00 GMT
to_http_date(Sys.time())

# parse HTTP date string to POSIXct
from_http_date("Thu, 01 Jan 1970 00:00:00 GMT") # .POSIXct(0, tz = "GMT")
from_http_date("Wed, 21 Aug 2019 15:36:48 GMT")

}
\references{
\href{https://tools.ietf.org/html/rfc7231#section-7.1.1.1}{RFC7231}
\href{https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Date}{MDN}
}
