% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{http_headers_to_split_default}
\alias{http_headers_to_split_default}
\title{Returns a vector of http header names which are split by default}
\usage{
http_headers_to_split_default()
}
\value{
A vector of http header names
}
\description{
Returns a vector of http header names which are split by default
}
\seealso{
\url{https://en.wikipedia.org/wiki/List_of_HTTP_header_fields} and
\url{https://stackoverflow.com/a/29550711/3048453}
}
