% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{http_statuses}
\alias{http_statuses}
\alias{http_404_not_found}
\alias{http_405_method_not_allowed}
\alias{http_500_internal_server_error}
\alias{http_520_unknown_error}
\title{Create standart responses/exceptions}
\usage{
http_404_not_found(body = "Page not found", content_type = "text/plain",
  ...)

http_405_method_not_allowed(body = "Method Not Allowed",
  content_type = "text/plain", ...)

http_500_internal_server_error(body = "Internal Server Error",
  content_type = "text/plain", ...)

http_520_unknown_error(body = "Unknown Error", content_type = "text/plain",
  ...)
}
\arguments{
\item{body}{reponse body}

\item{content_type}{body content type}

\item{...}{other parameters to \link{create_response}}
}
\description{
functions for convenient creation of standard http answers/exceptions
}
