% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{IDE-hints}
\alias{IDE-hints}
\alias{.req}
\alias{.res}
\title{request and reponse placeholders for IDE hints}
\format{
An object of class \code{Request} (inherits from \code{R6}) of length 28.

An object of class \code{Response} (inherits from \code{R6}) of length 26.
}
\usage{
.req

.res
}
\description{
request and reponse placeholders for IDE hints
}
\examples{
library(RestRserve)

app = Application$new()

app$add_get("/foo", FUN = function(.req, .res) {
  # since .res is a dummy instance of Response class
  # exported by RestRserve
  # IDE facilitates with autocompletion!
  .res$set_body("bar")
  # in the same time all the modifications happen with local objects
  # so you get right results in the end
})

response = app$process_request(Request$new(path = "/foo"))
response$body
}
\seealso{
\link{Request} \link{Response}
}
\keyword{datasets}
