% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\docType{data}
\name{logging_constants}
\alias{logging_constants}
\alias{OFF}
\alias{FATAL}
\alias{ERROR}
\alias{WARN}
\alias{INFO}
\alias{DEBUG}
\alias{TRACE}
\title{log level constants}
\format{An object of class \code{integer} of length 1.}
\usage{
OFF

FATAL

ERROR

WARN

INFO

DEBUG

TRACE
}
\description{
log level constants
}
\keyword{datasets}
