% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openapi.R
\name{openapi_create}
\alias{openapi_create}
\alias{openapi_openapi_version}
\alias{openapi_info}
\alias{openapi_servers}
\alias{openapi_server}
\alias{openapi_contact}
\alias{openapi_license}
\title{creates OpenAPI objects}
\usage{
openapi_create(openapi = openapi_openapi_version(), info = openapi_info(),
  servers = openapi_servers(), ...)

openapi_openapi_version(openapi_version = "3.0.1")

openapi_info(title = "RestRserve OpenAPI", version = "1.0",
  description = NULL, termsOfService = NULL, contact = openapi_contact(),
  license = openapi_license())

openapi_servers(servers = list())

openapi_server(url = "/", description = NULL, variables = NULL)

openapi_contact(name = NULL, url = NULL, email = NULL)

openapi_license(name = NULL, url = NULL)
}
\description{
Facilitates in building \href{https://www.openapis.org/}{OpenAPI} description document by
creating objects described in
 \url{https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.1.md}
}
