% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestRserveRequest.R
\name{parse_request}
\alias{parse_request}
\title{parses http request from Rserve}
\usage{
parse_request(path, query, body, headers)
}
\value{
\describe{
      \item{path}{ = \code{"/somepath"}, always character of length 1}
      \item{method}{ = \code{"GET"}, always character of length 1}
      \item{query}{ = \code{c("a" = "1", "b" = "2")}, named character vector. Query parameters key-value pairs.}
      \item{body}{ = \code{NULL}.
         \itemize{
            \item \code{NULL} if the http body is empty or zero length.
            \item \code{raw vector} with a "content-type" attribute in all cases except URL encoded form (if specified in the headers)
            \item named \code{characeter vector} in the case of a URL encoded form.
            It will have the same shape as the query string (named string vector).
         }
      }
      \item{content_type}{ = \code{""}, always character of length 1}
      \item{headers}{ = \code{c("a" = "1", "b" = "2")}, named character vector. key-value pairs from http-header.}
   }
}
\description{
internal (not part of public API) function for convenient parsing of
http request objects from Rserve backend.
}
