% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTTPError.R
\name{raise}
\alias{raise}
\title{interrupts request handling}
\usage{
raise(x)
}
\arguments{
\item{x}{instance of \code{RestRserveResponse}. Can be created using \link{HTTPError} -
see examples.}
}
\description{
interrupts request handling and signals RestRserve to return HTTPError
}
\examples{
caught_http_exception = try(raise(HTTPError$bad_request()), silent = TRUE)
condition = attr(caught_http_exception, 'condition')
# response is a valid RestRserveResponse instace
identical(condition$response$body, "400 Bad Request")
}
