% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{restrserve_defaul_conf}
\alias{restrserve_defaul_conf}
\title{Default RestRserve configuration}
\usage{
restrserve_defaul_conf()
}
\value{
named character vector - names are Rserve configuration parameters
and values are corresponding configuration entries
}
\description{
returns default RestRserve configuration
}
