% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop.R
\name{restrserve_stop}
\alias{restrserve_stop}
\title{stops RestRserve application}
\usage{
restrserve_stop(dir)
}
\arguments{
\item{dir}{character, path to the directory where application was deployed}
}
\value{
invisibly returns the status of the attempt to kill process group
}
\description{
Assuming that RserveApplication is deployed to a \code{dir},
\code{restrserve_stop} stops Rserve by executing \code{kill -INT -- -$PGID}
where $PGID is Rserve PID read from configured pid file. PID obtained from
information from \code{Rserve.conf} file inside \code{dir}.
}
