% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestRserveApplication.R
\docType{data}
\name{RestRserveApplication}
\alias{RestRserveApplication}
\title{Creates RestRserveApplication.}
\format{\code{\link{R6Class}} object.}
\usage{
RestRserveApplication
}
\description{
Creates RestRserveApplication object.
RestRserveApplication converts user-supplied R code into high-performance REST API by
allowing to easily register R functions for handling http-requests.
}
\section{Usage}{

\itemize{
\item \code{app = RestRserveApplication$new()}
\item \code{app$add_route(path = "/echo", method = "GET", FUN =  function(request, response) {
  response$body = request$query[[1]]
  response$content_type = "text/plain"
  forward()
  })}
\item \code{app$routes()}
}
For usage details see \bold{Methods, Arguments and Examples} sections.
}

\section{Arguments}{

\describe{
 \item{app}{A \code{RestRserveApplication} object}
 \item{path}{\code{character} of length 1. Should be valid path for example \code{'/a/b/c'}.
 If it is named character vector with name equal to \code{"prefix"} then all the endopoints which
 begin with the path will call corresponding handler.}
 \item{method}{\code{character} of length 1. At the moment one of
   \code{("GET", "HEAD", "POST", "PUT", "DELETE", "OPTIONS", "PATCH")}}
 \item{FUN}{\code{function} which \strong{must take 2 arguments - \code{request}, \code{response} objects}.
   See \link{RestRserveRequest} and \link{RestRserveResponse} for details.
 }
}
}

\keyword{datasets}
