% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestRserveAuthMiddleware.R
\docType{data}
\name{RestRserveAuthMiddleware}
\alias{RestRserveAuthMiddleware}
\title{Authorization middleware}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RestRserveAuthMiddleware
}
\description{
adds verious authorizations to \link{RestRserveApplication}
}
\keyword{datasets}
