% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestRserveMethodHandler.R
\docType{data}
\name{RestRserveMethodHandler}
\alias{RestRserveMethodHandler}
\title{Creates RestRserveMethodHandler.}
\format{\code{\link{R6Class}} object.}
\usage{
RestRserveMethodHandler
}
\description{
Creates RestRserveMethodHandler object.
}
\keyword{datasets}
