% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestRserveResponse.R
\docType{data}
\name{RestRserveResponse}
\alias{RestRserveResponse}
\title{Creates RestRserveResponse object (R6 class)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RestRserveResponse
}
\description{
Creates RestRserveResponse object.
Useful if handler or middleware wants to return some value immediately and interrupt request-response cycle.
\itemize{
\item \code{response = RestRserveResponse$new(body = "{}", content_type = "application/json", headers = character(0), status_code = 200L)}
\describe{
  \item{body}{must be a character vector of length one or a raw vector.
  If it is a named character with a name \code{file} or \code{tmpfile}
  then the value is considered as a path to a file and content oh this file is served as body.
  The latter will be deleted once served.}
  \item{content_type}{\code{"application/json"} must be a character vector of length one}
  \item{headers}{\code{character(0)} must be a character vector - the elements will have CRLF appended.
  Neither Content-type nor Content-length may be used.}
  \item{status_code}{\code{200L} must be an integer}
}
}
}
\section{Methods}{

\describe{
  \item{\code{$new(body = "{}", content_type = "application/json", headers = character(0), status_code = 200L)}}{Constructor for RestRserveResponse}
}
}

\keyword{datasets}
