% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestRserveRouter.R
\docType{data}
\name{RestRserveRouter}
\alias{RestRserveRouter}
\title{Creates RestRserveRouter.}
\format{\code{\link{R6Class}} object.}
\usage{
RestRserveRouter
}
\description{
Creates RestRserveRouter object.
}
\keyword{datasets}
