% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackendRserve.R
\name{ApplicationProcess}
\alias{ApplicationProcess}
\title{Creates ApplicationProcess object}
\description{
Creates ApplicationProcess to hold PID of the runnung applicaiton.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pid}}{Process identificator.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ApplicationProcess-new}{\code{ApplicationProcess$new()}}
\item \href{#method-ApplicationProcess-kill}{\code{ApplicationProcess$kill()}}
\item \href{#method-ApplicationProcess-clone}{\code{ApplicationProcess$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationProcess-new"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationProcess-new}{}}}
\subsection{Method \code{new()}}{
Creates ApplicationProcess object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationProcess$new(pid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pid}}{Process identificator.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationProcess-kill"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationProcess-kill}{}}}
\subsection{Method \code{kill()}}{
Send signal to process.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationProcess$kill(signal = 15L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signal}}{Singal code.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ApplicationProcess-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ApplicationProcess-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ApplicationProcess$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
