% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestRserveAuthMiddleware.R
\docType{data}
\name{AuthBackendBasic}
\alias{AuthBackendBasic}
\alias{BasicAuthBackend}
\title{Basic authorization backend}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AuthBackendBasic

BasicAuthBackend
}
\description{
\url{https://en.wikipedia.org/wiki/Basic_access_authentication}
}
\keyword{datasets}
