% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthMiddleware.R
\name{AuthMiddleware}
\alias{AuthMiddleware}
\title{Creates authorization middleware object}
\description{
Adds various authorizations to \link{Application}.
}
\seealso{
\link{Middleware} \link{Application}

Other AuthBackend: 
\code{\link{AuthBackend}},
\code{\link{AuthBackendBasic}},
\code{\link{AuthBackendBearer}}
}
\concept{AuthBackend}
\section{Super class}{
\code{\link[RestRserve:Middleware]{RestRserve::Middleware}} -> \code{AuthMiddleware}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AuthMiddleware-new}{\code{AuthMiddleware$new()}}
\item \href{#method-AuthMiddleware-clone}{\code{AuthMiddleware$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthMiddleware-new"></a>}}
\if{latex}{\out{\hypertarget{method-AuthMiddleware-new}{}}}
\subsection{Method \code{new()}}{
Creeates AuthMiddleware object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthMiddleware$new(
  auth_backend,
  routes,
  match = "exact",
  id = "AuthMiddleware"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{auth_backend}}{Authentication backend.}

\item{\code{routes}}{Routes paths to protect.}

\item{\code{match}}{How routes will be matched: \code{"exact"} or \code{"partial"} (as prefix).}

\item{\code{id}}{Middleware id.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthMiddleware-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AuthMiddleware-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthMiddleware$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
