% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestRserveContentHandlers.R
\docType{data}
\name{ContentHandlers}
\alias{ContentHandlers}
\title{ContentHandlers}
\format{An object of class \code{RestRserveContentHandler} (inherits from \code{R6}) of length 9.}
\usage{
ContentHandlers
}
\description{
ContentHandlers
}
\keyword{datasets}
