% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{create_response}
\alias{create_response}
\title{creates http response}
\usage{
create_response(body = "", content_type = "text/html",
  headers = character(0), status_code = 200L)
}
\arguments{
\item{body}{must be a character vector of length one or a raw vector.
If it is a named character with a name \code{file} or \code{tmpfile}
then the value is considered as a path to a file and content oh this file is served as body.
The latter will be deleted once served.}

\item{content_type}{\code{"text/html"} must be a character vector of length one}

\item{headers}{\code{character(0)} must be a character vector - the elements will have CRLF appended.
Neither Content-type nor Content-length may be used.}

\item{status_code}{\code{200L} must be an integer}
}
\value{
object of the class \code{"RestRserveResponse"} which is essentially a R's list.
}
\description{
facilitates in creation of proper response object. User functions should always return
object created with this function.
}
