% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EncodeDecodeMiddleware.R
\docType{data}
\name{EncodeDecodeMiddleware}
\alias{EncodeDecodeMiddleware}
\title{Creates EncodeDecodeMiddleware middleware object}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Controls how RestRserve encodes and decodes different content types.
\strong{This middleware is passed by default to the \link{Application} constructor}.
This class inherits from \link{Middleware}.
}
\section{Construction}{
\preformatted{EncodeDecodeMiddleware$new(id = "EncodeDecodeMiddleware")
}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Middleware id
}
}

\section{Fields}{

\itemize{
\item \strong{\code{ContentHandlers}} :: \code{ContentHandlers}\cr
Class which controls how RestRserve encodes and decodes different content types.
See \link{ContentHandlers} for documentation.
User can add new ecoding and decoding methods for new content types using \code{set_encode}
and \code{set_decode} methods.

In theory user can replace it with his own class (see \code{RestRserve:::ContentHandlersFactory}).
However we believe that in the majority of the cases using \link{ContentHandlers} will be enough.
}
}

\seealso{
\link{Middleware} \link{Application} \link{ContentHandlers}
}
\keyword{datasets}
