% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTTPError.R
\docType{data}
\name{HTTPError}
\alias{HTTPError}
\title{HTTPError}
\format{An object of class \code{HTTPErrorFactory} (inherits from \code{R6}) of length 38.}
\usage{
HTTPError
}
\description{
Contains \link{HTTPErrorFactory} class for the exception in the user's code.
}
\seealso{
\link{raise} \link{HTTPErrorFactory}
}
\keyword{datasets}
