% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTTPError.R
\docType{data}
\name{HTTPErrorFactory}
\alias{HTTPErrorFactory}
\title{helps to generate http error responces}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HTTPErrorFactory
}
\description{
helps to generate http error responces. See \link{raise} for example.
}
\section{Methods}{

\describe{
\item{\code{$new(content_type = "text/plain", serializer = NULL)}}{Factory constructor.
\describe{
\item{content_type}{type of the error response. \code{"text/plain"} by default}
\item{serializer}{\code{NULL} (default) or function. Specify how encode response body. If \code{NULL}
then RestRserve will try to automatically encode body properly according to \code{content_type} argument}
}
}
}
}

\keyword{datasets}
