% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\docType{data}
\name{Logger}
\alias{Logger}
\title{simple logging utility}
\format{\code{\link{R6Class}} object.}
\usage{
Logger
}
\description{
creates Logger object which can be used for logging with different level of verbosity.
Log messages are in JSON format
}
\section{Methods}{

\describe{
\item{\code{$new(level = INFO, name = "ROOT", printer = NULL)}}{Logger with sink defined by \code{printer}
function. It should have signature \code{function(timestamp, level, logger_name, pid, message)}.
By default when \code{printer = NULL} logger writes message in JSON format to \code{stdout}}
\item{\code{$set_name(name = "ROOT")}}{ sets logger name}
\item{\code{$set_log_level(level = INFO)}}{ sets log level}
\item{\code{$set_printer(FUN = NULL)}}{ sets function which defines how to print logs}
\item{\code{$trace(msg, ...)}}{ write trace message}
\item{\code{$debug(msg, ...)}}{ write debug message}
\item{\code{$info(msg, ...)}}{ write info message}
\item{\code{$warning(msg, ...)}}{ write warning message}
\item{\code{$error(msg, ...)}}{ write error message}
}
}

\examples{
logger = Logger$new(INFO)
logger$info("hello world")
logger$info(list(message = "hello world", code = 0L))
}
\keyword{datasets}
