% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\docType{data}
\name{Logger}
\alias{Logger}
\title{simple logging utility}
\format{\code{\link{R6Class}} object.}
\usage{
Logger
}
\description{
creates Logger object which can be used for logging with different level of verbosity
}
\section{Methods}{

\describe{
  \item{\code{$new(level = INFO, file = "")}}{Logger with sink to a \code{file}. \code{file} can be character
  or connection. Internally passed to \link{cat} - see corresponding docs for details.}
  \item{\code{$trace(msg, ...)}}{ write trace message}
  \item{\code{$debug(msg, ...)}}{ write debug message}
  \item{\code{$info(msg, ...)}}{ write info message}
  \item{\code{$warning(msg, ...)}}{ write warning message}
  \item{\code{$error(msg, ...)}}{ write error message}
}
}

\examples{
logger = Logger$new(INFO)
logger$info("hello world")
logger$info(list(message = "hello world", code = 0L))
}
\keyword{datasets}
