% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTTPError.R
\name{raise}
\alias{raise}
\title{interrupts request handling}
\usage{
raise(x)
}
\arguments{
\item{x}{instance of \code{HTTPError}. Can be created using \link{HTTPErrorFactory} -
see examples.}
}
\description{
interrupts request handling and signals RestRserve to return HTTPError
}
\examples{
err_factory = HTTPErrorFactory$new()
caught_http_exception = try(raise(err_factory$bad_request()), silent = TRUE)
condition = attr(caught_http_exception, 'condition')
# response is an instance of HTTPError
# and valid RestRserveResponse instace because HTTPError inherits from RestRserveResponse
identical(condition$response$body, "400 Bad Request")
}
