% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTTPError.R
\name{raise}
\alias{raise}
\title{interrupts request handling}
\usage{
raise(x)
}
\arguments{
\item{x}{instance of \link{RestRserveResponse}. Can be created using \link{HTTPError}.
see examples.}
}
\description{
Interrupts request handling and signals RestRserve to return HTTPError
}
\examples{
# catch exception
res = try(raise(HTTPError$bad_request()), silent = TRUE)
cond = attr(res, "condition")

# response is a valid RestRserveResponse instace
identical(cond$response$body, "400 Bad Request")

}
\seealso{
\link{HTTPError}
}
