% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{restrserve_deploy}
\alias{restrserve_deploy}
\title{Deploys RestRserve application to a directory}
\usage{
restrserve_deploy(file, dir = "RestRserveApplication",
  configuration = restrserve_default_conf(), configuration_file = NULL,
  pid_file = file.path(dir, "Rserve.pid"))
}
\arguments{
\item{file}{file with user code. In order to deploy application user have to
creates instance of \code{RestRserveApplication} \bold{with a name \code{RestRserveApp}}.}

\item{dir}{'\code{character}, path to the directory where to deploy application and configuration}

\item{configuration}{character vector - names are Rserve configuration parameters
and values are corresponding configuration entries}

\item{configuration_file}{\code{NULL} or path to a file with Rserve configuration. If not \code{NULL} then
values of the \code{configuration_file} will be added after parameter values
from \code{configuration} argument above.}

\item{pid_file}{character, path to a file where to put PID after starting application
(after \link{restrserve_start} call).}
}
\value{
\code{TRUE} invisibly if deployment was successful
}
\description{
Takes file with RestRserve application, generates Rserve configuration
and put it to a specified directory. Later application could be started with \link{restrserve_start}.
}
