% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestRserveAuthMiddleware.R
\docType{data}
\name{RestRserveAuthMiddleware}
\alias{RestRserveAuthMiddleware}
\title{Creates authorization middleware object}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Adds various authorizations to \link{RestRserveApplication}.
This class inherits \link{RestRserveMiddleware}.
}
\section{Construction}{
\preformatted{RestRserveAuthMiddleware$new(auth_backend, routes, match = "exact",
                             name = "AuthMiddleware")
}
\itemize{
\item \code{auth_backend} :: \link{AuthBackend}\cr
Authentication backend.
\item \code{routes} :: \code{character()}\cr
Routes paths to protect.
\item \code{match} :: \code{character()}\cr
How routes will be matched: exact or partial (as prefix).
\item \code{name} :: \code{character(1)}\cr
Middleware name.
}
}

\seealso{
\link{RestRserveMiddleware} \link{RestRserveApplication}

Other AuthBackend: \code{\link{AuthBackendBasic}},
  \code{\link{AuthBackendBearer}},
  \code{\link{AuthBackend}}
}
\concept{AuthBackend}
\keyword{datasets}
