% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestRserveRequest.R
\docType{data}
\name{RestRserveRequest}
\alias{RestRserveRequest}
\title{Creates RestRserveRequest object (R6 class)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RestRserveRequest
}
\value{
\code{RestRserveRequest} object - R6 class:
\describe{
\item{path}{ = \code{"/somepath"}, always character of length 1}
\item{method}{ = \code{"GET"}, always character of length 1}
\item{query}{\code{list("a" = "1", "b" = "2")}, \bold{list}, key-value pairs from query parameters.}
\item{body}{ = \code{raw(0)}.
\itemize{
\item \code{NULL} if the http body is empty or zero length.
\item \code{raw vector} in all cases except URL encoded form
\item named \code{characeter vector} in the case of a URL encoded form.
It will have the same shape as the query string (named string vector).
}
},
\item{body_decoded}{ body parsed according to the 'content-type' request header
and \code{decode} argument of the r
}
\item{request_id}{\bold{character}, automatically generated UUID for each request}
\item{content_type}{\code{""}, always character of length 1}
\item{headers}{ \code{list("a" = "1", "b" = "2")}, \bold{list}, key-value pairs from http-header.
According to \href{https://www.w3.org/Protocols/rfc2616/rfc2616-sec4.html#sec4.2}{RFC2616} header field names
are case-insensitive. So in RestRserve \bold{keys are always in lower case}.
},
\item{path_parameters}{\bold{list()}, list of parameters extracted from templated path after routing.
For example if we have some hadler listening at \code{/job/{job_id}}} and we are receiving request at
\code{/job/1} then \code{path_parameters} will be \code{list(job_id = "1")}. It is important to understand
that \code{path_parameters} will be available (not empty) only after request will reach handler. This
effectively means that \code{path_parameters} can be used inside handler and response middleware
(but not request middleware!)
}
}
\description{
Called internally for handling incoming requests from Rserve side. Also useful for testing.
\itemize{
\item \code{response = RestRserveRequest$new(
  path = "/",
  method = "GET",
  query = new.env(parent = emptyenv()),
  headers = new.env(parent = emptyenv()),
  body = raw(),
  decode = NULL)}
\describe{
\item{path}{\code{"/somepath"}, always character of length 1}
\item{method}{\code{"GET"}, always character of length 1}
\item{query}{\code{list("a" = "1", "b" = "2")}, \bold{list}, key-value pairs from query parameters.}
\item{body}{
\itemize{
\item \code{NULL} if the http body is empty or zero length.
\item \code{raw vector} with a "content-type" attribute in all cases except URL encoded form (if specified in the headers)
\item named \code{characeter vector} in the case of a URL encoded form.
It will have the same shape as the query string (named string vector)}
}
\item{headers}{ \code{list("a" = "1", "b" = "2")}, \bold{list}, key-value pairs from http-header.}
}
}
}
\keyword{datasets}
