% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{to_json}
\alias{to_json}
\title{simple json encoder}
\usage{
to_json(x, unbox = TRUE)
}
\arguments{
\item{x}{an R object. Lists (possibly named and nested) and arrays are supported}

\item{unbox}{whether to unbox (simplify) arrays consists of a single element}
}
\description{
encode R objects as json
}
