% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{to_json}
\alias{to_json}
\title{Simple JSON encoder}
\usage{
to_json(x, unbox = TRUE)
}
\arguments{
\item{x}{the object to be encoded}

\item{unbox}{\code{TRUE} by default. Whether to unbox (simplify) arrays consists
of a single element}
}
\value{
JSON string
}
\description{
Encode R objects as JSON. Wrapper around \code{jsonlite::toJSON} with
default parameters set to following values:
\verb{dataframe = 'columns', auto_unbox = unbox, null = 'null', na = 'null'}.
}
\examples{
to_json(NULL)
to_json(list(name = "value"))

}
